refactor(frontend_lib): Moved every columns and dependencies to frontend_lib
