+++ frontend/src/container/tables/Tables/FavoriteTable/FavoriteTable.styl
