+++ frontend_lib/src/component/ContentType/ContentType.styl
@@ -1,4 +1,4 @@
-@import "~tracim_frontend_lib/src/css/Variable.styl"
+@import "../../css/Variable.styl"
 
 .contentType
   font-size titleFontSize
